requirements.txt
appnope==0.1.0
certifi==2017.11.5
chardet==3.0.4
decorator==4.1.2
Django==2.0.1
django-appconf==1.0.2
django-extensions==1.9.9
django-imagekit==4.0.2
idna==2.6
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.1
parso==0.1.1
pexpect==4.3.1
pickleshare==0.7.4
pilkit==2.0
Pillow==5.0.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
pytz==2017.3
requests==2.18.4
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
typing==3.6.2
urllib3==1.22
wcwidth==0.1.7
